Tax_Assistant/requirements.txt
flask
flask-cors
google-generativeai

